Refactor dependency logic for publish modal
Extracted resource dependency logic into a new useResourceDependencies hook for improved modularity and reusability in the dashboard. Optimized dependency analysis in the backend by batching workflow control value queries. Also introduced a constant for publish batch size in the publish environment use case.
